requirements.txt
asgiref==3.6.0
Django==4.1.5
django-autoslug==1.9.8
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
